% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztheme.R
\name{apply_z_theme}
\alias{apply_z_theme}
\title{Apply Z Theme Colors provided by Zoni Nation -> Perceptions}
\usage{
apply_z_theme()
}
\value{
A theme object.
}
\description{
Apply Z Theme Colors provided by Zoni Nation -> Perceptions available at
https://github.com/zonination/perceptions. The code is inspired by ztheme.R
file available within perceptions package at
https://github.com/zonination/perceptions/blob/master/ztheme.R.
}
\details{
apply_z_theme
}
\examples{
apply_z_theme()
}
\author{
Zoni Nation - https://github.com/zonination
}
