% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahitya_akademi_data.R
\name{get_author_html}
\alias{get_author_html}
\title{Get Author HTML Content from WikiPedia}
\usage{
get_author_html(author)
}
\arguments{
\item{author}{a character vector of length 1. For example, 'Ruskin Bond'.}
}
\value{
An xml_document containing the html text with author's details
like date of birth, etc.
}
\description{
Get Author HTML Content from WikiPedia.
}
\details{
get_author_html
}
\examples{
get_author_html("Ruskin Bond")
}
\author{
Pushker Ravindra
}
