% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahitya_akademi_data.R
\name{get_authors_dob_dod}
\alias{get_authors_dob_dod}
\title{Get Authors Date of Birth and Death}
\usage{
get_authors_dob_dod(language_data)
}
\arguments{
\item{language_data}{a data.frame containing list of Sahitya Akademi award
winners with the following information: Year, Author, Work, Category.}
}
\value{
A data.frame containing Year, Author, Work, Category, Date of Birth,
Date of Death, Remarks.
}
\description{
Get authors date of birth and death from the authors' Wikipedia
page.
}
\details{
get_authors_dob_dod
}
\examples{
language_data <- get_sahitya_akademi_winners("Hindi")[1:5, ]
get_authors_dob_dod(language_data)
}
\author{
Ravindra Pushker
}
