% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahitya_akademi_data.R
\name{get_dob_from_description}
\alias{get_dob_from_description}
\title{Get the Date of Birth of an Author from the Description}
\usage{
get_dob_from_description(author_html)
}
\arguments{
\item{author_html}{a xml_document containing the html text with author's
details like date of birth, etc.}
}
\value{
A list of data of birth and date of death with remarks in case there were
some challenges in getting date of birth / death.
}
\description{
Get the date of birth of an author from the description when the info box is
not present.
}
\details{
get_dob_from_description
}
\examples{
author_html <- get_author_html("Ruskin Bond")
get_dob_from_description(author_html)
}
\author{
Pushker Ravindra
}
