% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahitya_akademi_data.R
\name{get_dob_from_infobox}
\alias{get_dob_from_infobox}
\title{Get Date of Birth from the Author's InfoBox}
\usage{
get_dob_from_infobox(author_info_df)
}
\arguments{
\item{author_info_df}{a data.frame containing 2 colums Field and Value. Field
contains information like Born, Died, which is required to get date of birth.}
}
\value{
A list of data of birth and date of death with remarks in case there were
some challenges in getting date of birth / death.
}
\description{
Get Date of Birth from the Author's InfoBox which is on the top
right side of author's page below their image.
}
\details{
get_dob_from_infobox
}
\examples{
author_info_df <- data.frame(Field=c("Born", "Died"), Value=c("Harivansh Rai
Srivastava(1907-11-27)27 November 1907Babupatti, United Provinces of Agra and
Oudh, British India", "18 January 2003(2003-01-18) (aged 95)"))
get_dob_from_infobox(author_info_df)
}
\author{
Ravindra Pushker
}
