% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahitya_akademi_data.R
\name{get_sahitya_akademi_languages}
\alias{get_sahitya_akademi_languages}
\title{Get Sahitya Akademi Languages}
\usage{
get_sahitya_akademi_languages()
}
\value{
A character vector of length 24.
}
\description{
Sahitya Akademi Award is given in 24 languages. Return those 24 languages.
}
\details{
get_sahitya_akademi_languages
}
\examples{
get_sahitya_akademi_languages
}
\author{
Pushker Ravindra
}
