% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahitya_akademi_data.R
\name{get_sahitya_akademi_languages_data}
\alias{get_sahitya_akademi_languages_data}
\title{Get Sahitya Akademi Languages Data}
\usage{
get_sahitya_akademi_languages_data(languages)
}
\arguments{
\item{languages}{a character vector, for example, 'English'.}
}
\value{
A data.frame
}
\description{
Get Sahitya Akademi data for the given languages. For each language, the data is stored in a CSV file and for given language(s), the data is fetched and aggregated to provide data for the given languages.
}
\details{
get_sahitya_akademi_languages_data
}
\examples{
get_sahitya_akademi_languages_data(c("English", "Hindi", "Kannada"))
}
\author{
Pushker Ravindra
}
