% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahitya_akademi_data.R
\name{get_sahitya_akademi_winners}
\alias{get_sahitya_akademi_winners}
\title{Get Sahitya Akademi Winners for any Language}
\usage{
get_sahitya_akademi_winners(language)
}
\arguments{
\item{language}{a character vector of length 1, for exampple, 'Emglish'.}
}
\value{
A data.frame containing list of Sahitya Akademi award winners with
the following information: Year, Author, Work, Category.
}
\description{
Get Sahitya Akademi award winners from the Wikipedia page -
https://en.wikipedia.org/wiki/Lists_of_Sahitya_Akademi_Award_winners. For
each language there is a separate page, for example, for English, the details
are available on https://en.wikipedia.org/wiki/List_of_Sahitya_Akademi_Award_winners_for_English.
}
\details{
get_sahitya_akademi_winners
}
\examples{
get_sahitya_akademi_winners("Hindi")
}
\author{
Ravindra Pushker
}
