% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahitya_akademi_data.R
\name{is_sahitya_akademi_language}
\alias{is_sahitya_akademi_language}
\title{Is Sahitya Akademi Language}
\usage{
is_sahitya_akademi_language(languages)
}
\arguments{
\item{languages}{a character vector, for example, 'English'.}
}
\value{
TRUE or FALSE
}
\description{
Check whether Sahitya Akademi Award is given in the provided language(s). Sahitya Akademi Award is given in 24 languages.
}
\details{
is_sahitya_akademi_language
}
\examples{
is_sahitya_akademi_language(c("English", "Hindi", "Kannada"))
is_sahitya_akademi_language("Spanish")
}
\author{
Pushker Ravindra
}
