% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sahitya_akademi.R
\name{plot_sahitya_akademi_age}
\alias{plot_sahitya_akademi_age}
\title{Get the Date of Birth of an Author from the Description}
\usage{
plot_sahitya_akademi_age(languages, type = "box", ztheme = TRUE)
}
\arguments{
\item{languages}{a character vector containing any of the 24 languages.}

\item{type}{a character vector of length 1 for the type of plot whether box or violin.}

\item{ztheme}{a logical vector of length 1 for applying ztheme.}
}
\value{
A ggplot object containing box/violin plot.
}
\description{
Get the date of birth of an author from the description when the info box is
not present.
}
\details{
plot_sahitya_akademi_age
}
\examples{
plot_sahitya_akademi_age(c("Hindi", "English", "Kannada"))
}
\author{
Ravindra Pushker
}
